requirements/base.txt
Django==4.2.5
django-ckeditor==6.7.0
django-extensions==3.2.3
djangorestframework==3.14.0
mysqlclient==2.2.0
requests==2.31.0
typesense==0.17.0
